git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@15838 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickCore/cache.c
@@ -601,7 +601,8 @@ static MagickBooleanType ClonePixelCacheRepository(
       clone_nexus[id],exception);
     if (pixels == (Quantum *) NULL)
       continue;
-    (void) ResetMagickMemory(clone_nexus[id]->pixels,0,clone_nexus[id]->length);
+    (void) ResetMagickMemory(clone_nexus[id]->pixels,0,(size_t)
+      clone_nexus[id]->length);
     if (optimize != MagickFalse)
       (void) memcpy(clone_nexus[id]->pixels,cache_nexus[id]->pixels,length*
         sizeof(Quantum));
